CLDR-16239 Fixed remaining problems (pending review)

+++ tools/cldr-code/src/main/java/org/unicode/cldr/tool/CLDRModify.java
@@ -19,6 +19,7 @@ import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
+import java.util.Objects;
 import java.util.Set;
 import java.util.TreeMap;
 import java.util.TreeSet;
@@ -71,6 +72,7 @@ import org.unicode.cldr.util.XPathParts.Comments.CommentType;
 
 import com.google.common.base.Joiner;
 import com.google.common.base.Splitter;
+import com.google.common.collect.ImmutableList;
 import com.ibm.icu.dev.tool.UOption;
 import com.ibm.icu.impl.Utility;
 import com.ibm.icu.text.Collator;
@@ -230,7 +232,9 @@ public class CLDRModify {
         USER = 12,
         ALL_DIRS = 13,
         CHECK = 14,
-        KONFIG = 15;
+        KONFIG = 15,
+        RETAIN = 16
+        ;
 
     private static final UOption[] options = {
         UOption.HELP_H(),
@@ -249,6 +253,7 @@ public class CLDRModify {
         UOption.create("all", 'a', UOption.REQUIRES_ARG),
         UOption.create("check", 'c', UOption.NO_ARG),
         UOption.create("konfig", 'k', UOption.OPTIONAL_ARG).setDefault("modify_config.txt"),
+        UOption.create("Retain", 'R', UOption.NO_ARG),
     };
 
     private static final UnicodeSet allMergeOptions = new UnicodeSet("[rcd]");
@@ -300,7 +305,10 @@ public class CLDRModify {
         + "\t\thttp://cldr.unicode.org/development/cldr-big-red-switch/cldrmodify-passes/cldrmodify-config."
         + XPathParts.NEWLINE
         + "-f\t to perform various fixes on the files (add following arguments to specify which ones, eg -fxi)"
-        + XPathParts.NEWLINE;
+        + XPathParts.NEWLINE
+        + "-R\t retain unchanged files"
+        + XPathParts.NEWLINE
+        ;
 
     static final String HELP_TEXT2 = "Note: A set of bat files are also generated in <dest_dir>/diff. They will invoke a comparison program on the results."
         + XPathParts.NEWLINE;
@@ -334,6 +342,8 @@ public class CLDRModify {
         String targetDirBase = CldrUtility.checkValidDirectory(destInput); // Utility.GEN_DIRECTORY + "main/";
         System.out.format("Source:\t%s\n", sourceDirBase);
         System.out.format("Target:\t%s\n", targetDirBase);
+
+        boolean retainUnchangedFiles = options[RETAIN].doesOccur;
 
         Set<String> dirSet = new TreeSet<>();
         if (recurseOnDirectories == null) {
@@ -463,6 +473,7 @@ public class CLDRModify {
                     }
                     if (options[FIX].doesOccur) {
                         fix(k, options[FIX].value, options[KONFIG].value, cldrFactory);
+                        System.out.println("#TOTAL\tItems changed: " + fixList.totalChanged);
                     }
                     if (DEBUG_PATHS != null) {
                         System.out.println("Debug3 (" + test + "):\t" + k.toString(DEBUG_PATHS));
@@ -501,7 +512,8 @@ public class CLDRModify {
 
                     File oldFile = new File(sourceDir, test + ".xml");
                     File newFile = new File(targetDir, test + ".xml");
-                    if (!oldFile.equals(newFile) // only skip if the source & target are different.
+                    if (!retainUnchangedFiles
+                        && !oldFile.equals(newFile) // only skip if the source & target are different.
                         && equalsSkippingCopyright(oldFile, newFile)) {
                         newFile.delete();
                         continue;
@@ -625,6 +637,7 @@ public class CLDRModify {
         private Set<String> toBeRemoved;
         private CLDRFile toBeReplaced;
         protected Factory factory;
+        protected int countChanges;
 
         /**
          * Called when a new locale is being processed
@@ -640,6 +653,7 @@ public class CLDRModify {
             localeID = k.getLocaleID();
             this.toBeRemoved = removal;
             this.toBeReplaced = replacements;
+            countChanges = 0;
             handleStart();
         }
 
@@ -712,6 +726,7 @@ public class CLDRModify {
                 + "\t→\t" + (newValue == null ? "∅" : newValue.equals(oldValueOldPath) ? "≡" : "«" + newValue + "»")
                 + "\t" + oldFullPath
                 + (newFullPath.equals(oldFullPath) ? "" : "\t→\t" + newFullPath));
+            ++countChanges;
         }
 
         /**
@@ -817,6 +832,7 @@ public class CLDRModify {
         String[] helps = new String[128]; // only ascii
         UnicodeSet options = new UnicodeSet();
         String inputOptions = null;
+        int totalChanged = 0;
 
         void add(char letter, String help) {
             add(letter, help, null);
@@ -898,6 +914,10 @@ public class CLDRModify {
                 if (filters[c] != null) {
                     try {
                         filters[c].handleEnd();
+                        if (filters[c].countChanges != 0) {
+                            totalChanged += filters[c].countChanges;
+                            System.out.println("#" + filters[c].localeID + "\tItems changed: " + filters[c].countChanges);
+                        }
                     } catch (RuntimeException e) {
                         System.err.println("Failure in " + filters[c].localeID + "\t START");
                         throw e;
@@ -2434,31 +2454,142 @@ public class CLDRModify {
         });
 
         fixList.add('V', "Fix values that would inherit laterally", new CLDRFilter() {
-            boolean skip;
+            boolean skip = false;
+            boolean isL1 = false;
+            String parentId = null;
+            CLDRFile parentFile = null;
+            Set<String> pathsHandled = new HashSet<>();
+            String onlyValues = null;
+            String message = null;
+
             @Override
             public void handleStart() {
                 // skip if the locale id's parent isn't root. That is, it must be at level-1 locale.
-                skip = !XMLSource.ROOT_ID.equals(LocaleIDParser.getParent(getLocaleID()));
+                skip = getLocaleID().equals(XMLSource.ROOT_ID);
+                if (!skip) {
+                    parentId = LocaleIDParser.getParent(getLocaleID());
+                    isL1 = parentId.equals(XMLSource.ROOT_ID);
+                    parentFile = null; // lazy evaluate
+                }
+                pathsHandled.clear();
+                onlyValues = CldrUtility.INHERITANCE_MARKER;
+                message = "fix ↑↑↑ lateral";
             }
+
             @Override
             public void handlePath(String xpath) {
                 if (skip) {
                     return;
                 }
                 String value = cldrFileToFilter.getStringValue(xpath);
-                if (CldrUtility.INHERITANCE_MARKER.equals(value)) {
-                    Output<String> pathWhereFound = new Output<>();
-                    Output<String> localeWhereFound = new Output<>();
-                    String baileyValue = getResolved().getBaileyValue(xpath, pathWhereFound, localeWhereFound);
-                    if (baileyValue != null
-                        && !xpath.equals(pathWhereFound.value)
-                        && !GlossonymConstructor.PSEUDO_PATH.equals(pathWhereFound.value)) {
+                if (!Objects.equals(onlyValues, value)) {
+                    return;
+                }
+
+                // remember which paths we handle, so we can skip them in handleEnd
+                pathsHandled.add(xpath);
+
+                Output<String> pathWhereFound = new Output<>();
+                Output<String> localeWhereFound = new Output<>();
+                String baileyValue = getResolved().getBaileyValue(xpath, pathWhereFound, localeWhereFound);
+                if (baileyValue != null
+                    && !xpath.equals(pathWhereFound.value)
+                    && !GlossonymConstructor.PSEUDO_PATH.equals(pathWhereFound.value)) {
+
+                    // we have lateral inheritance, so we decide whether to harden.
+
+                    boolean harden = false;
+                    String message2 = "";
+
+                    // if we are L1, then we make a hard value, to protect higher values
+
+                    if (isL1) {
+                        harden = true;
+                        message2 = "; L1";
+                    } else {
+                        // for all others, we check to see if the parent's lateral value is the same as ours
+                        // If it is, we are ok, since one of that parent's parents will be hardened
+
+                        if (parentFile == null) {
+                            parentFile = factory.make(parentId, true);
+                        }
+                        String parentValue = parentFile.getStringValueWithBailey(xpath);
+                        if (!parentValue.equals(baileyValue)) {
+                            harden = true;
+                        }
+                        message2 = "; L2+";
+
+                        // Problem case: the parent value is null (not inheritance marker)
+                        // but the child value is ^^^.
+                        // See if we need to fix that.
+                    }
+                    if (harden) {
                         String fullPath = cldrFileToFilter.getFullXPath(xpath);
-                        replace(fullPath, fullPath, baileyValue, "fix lateral");
+                        replace(fullPath, fullPath, baileyValue, message + message2);
+                    }
+                }
+            }
+
+            @Override
+            public void handleEnd() {
+                if (skip || isL1) {
+                    return;
+                }
+                // Handle all the null cases that are in the L1 value.
+                onlyValues = null;
+                message = "fix null lateral";
+
+                List<String> parentChain = getParentChain(getLocaleID());
+                String localeL1 = parentChain.get(parentChain.size() - 2); // get last before root
+                CLDRFile fileL1 = factory.make(localeL1, false); // only unresolved paths
+                for (String path : fileL1) {
+                    if (!pathsHandled.contains(path)) {
+                        handlePath(path);
                     }
                 }
             }
         });
+    }
+
+    // TODO Move to central location
+
+    static final ImmutableList<String> FALLBACK_CHAIN = ImmutableList.of();
+    static final ImmutableList<String> ROOT_PARENT_CHAIN = ImmutableList.of(XMLSource.ROOT_ID);
+
+    /**
+     * Return the L1 locale in localeIds getParent chain.
+     * Return null if there is none (localeID == root or code-fallback).
+     * TODO optimize by caching the chains
+     * Returns a
+     */
+    public static List<String> getParentChain(String localeID) {
+        if (XMLSource.ROOT_ID.equals(localeID)) {
+            return FALLBACK_CHAIN;
+        }
+        List<String> result = null;
+        while (true) {
+            String parent = LocaleIDParser.getParent(localeID);
+            if (parent.equals(XMLSource.ROOT_ID)) {
+                if (result == null) {
+                    return ROOT_PARENT_CHAIN;
+                } else {
+                    result.addAll(ROOT_PARENT_CHAIN);
+                    return ImmutableList.copyOf(result);
+                }
+            }
+            if (result == null) {
+                result = new ArrayList<>();
+            }
+            result.add(parent);
+            localeID = parent;
+        }
+    }
+
+    static {
+        System.out.println("en_DE" + "\t" + getParentChain("en_DE"));
+        System.out.println("fr_CA" + "\t" + getParentChain("fr_CA"));
+        System.out.println("fr" + "\t" + getParentChain("fr"));
+        System.out.println("root" + "\t" + getParentChain("root"));
     }
 
     public static String getLast2Dirs(File sourceDir1) {
@@ -2548,6 +2679,7 @@ public class CLDRModify {
             fixList.handlePath(xpath);
         }
         fixList.handleEnd();
+
 
         // remove bad attributes
